+++ src/Tabs.php
@@ -9,6 +9,7 @@ use Eminiarts\Tabs\Contracts\TabContract;
 use Illuminate\Http\Resources\MergeValue;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Str;
 use Laravel\Nova\Contracts\ListableField;
 use Laravel\Nova\Panel;
 use Laravel\Nova\ResourceToolElement;
@@ -25,19 +26,23 @@ class Tabs extends Panel
     /**
      * @var bool
      */
-    public $showTitle = false;
+    public bool $showTitle = false;
 
     /**
      * @var bool
      */
-    public $selectFirstTab = true;
+    public bool $selectFirstTab = true;
 
-    /**
-     * @var TabContract[]
-     */
-    private $tabs = [];
+    private int $tabsCount = 0;
 
-    private $tabsCount = 0;
+    public $slug = null;
+
+    private string $preservedName;
+
+    public bool $retainTabPosition = false;
+
+    public $currentColor = null;
+    public string $errorColor = 'red';
 
     /**
      * Create a new panel instance.
@@ -49,9 +54,65 @@ class Tabs extends Panel
     public function __construct($name, $fields = [])
     {
         $this->name = $name;
+        $this->preservedName = $name;
         $this->withComponent('tabs');
 
         parent::__construct($this->prepareFields($fields));
+    }
+
+
+    /**
+     * Set the tabs slug.
+     *
+     * @param  string|boolean $slug
+     * @return $this
+     */
+    public function withSlug($slug): Tabs
+    {
+
+        $this->slug = is_bool($slug) ? ($slug ? Str::slug($this->preservedName, '_') : null) : $slug;
+
+        return $this;
+    }
+
+    /**
+     * Set the color for current tabs.
+     *
+     * @param string $color
+     * @return $this
+     */
+    public function withCurrentColor(string $color): Tabs
+    {
+
+        $this->currentColor = $color;
+
+        return $this;
+    }
+
+    /**
+     * Set the color for tabs with errors.
+     *
+     * @param  string $color
+     * @return $this
+     */
+    public function withErrorColor(string $color): Tabs
+    {
+
+        $this->errorColor = $color;
+
+        return $this;
+    }
+
+    /**
+     * Remember tab position across detail/edit
+     *
+     * @param  Boolean $retain
+     * @return $this
+     */
+    public function rememberTabs($retain): Tabs
+    {
+        $this->retainTabPosition = $retain;
+        return $this;
     }
 
     /**
@@ -60,7 +121,7 @@ class Tabs extends Panel
      * @param  Closure|array  $fields
      * @return array
      */
-    protected function prepareFields($fields)
+    protected function prepareFields($fields): array
     {
         $this->convertFieldsToTabs($fields)
             ->filter(static function (Tab $tab): bool {
@@ -217,6 +278,10 @@ class Tabs extends Panel
         $result = array_merge(parent::jsonSerialize(), [
             'defaultSearch' => $this->defaultSearch,
             'showTitle' => $this->showTitle,
+            'slug' => $this->slug,
+            'retainTabPosition' => $this->retainTabPosition,
+            'currentColor' => $this->currentColor,
+            'errorColor' => $this->errorColor
         ]);
 
         return $result;